pyproject.toml
[project]
name = "ai-cli-agent"
version = "0.1.0"
description = "Basic cli agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
